+++ models/item.js
@@ -1,19 +1,17 @@
 const Sequelize = require('sequelize');
 const sequelize = require('./sequelize.js');
+const Order = sequelize.import('../models/order.js');
 
- Item = sequelize.define('item', {
-    
-        itemName: Sequelize.STRING,
-        itemPrice: Sequelize.DOUBLE,
-        ingredient: Sequelize.STRING,
-        description: Sequelize.STRING,
-        rating: Sequelize.INTEGER,
-        foodID: Sequelize.DOUBLE,
-        itemQuantity: Sequelize.INTEGER
-    })
-	
+Item = sequelize.define('item', {
+  itemName: Sequelize.STRING,
+  itemPrice: Sequelize.DOUBLE,
+  ingredient: Sequelize.STRING,
+  description: Sequelize.STRING,
+  rating: Sequelize.INTEGER,
+  foodID: Sequelize.DOUBLE,
+  itemQuantity: Sequelize.INTEGER
+})
+Item.belongsToMany(Order, { through: 'OrderItem' });
 module.exports = (sequelize, DataTypes) => {
   return Item;
 }
-item.belongsToMany(ingredient);
-ingredient.belongsToMany(item);
