+++ models/order.js
@@ -1,16 +1,13 @@
 const Sequelize = require('sequelize');
-const sequelize = require('./sequelizeConf.js');
+const sequelize = require('./sequelize.js');
+const Item = sequelize.import('../models/item.js');
 
-Order = sequelize.define('order',
-    {
-        totalPrice: Sequelize.DOUBLE,
-        orderedItems: Sequelize.STRING,
-        specialReuest: Sequelize.STRING
-    })
+Order = sequelize.define('order', {
+  totalPrice: Sequelize.DOUBLE,
+  specialReuest: Sequelize.STRING
+})
 
+Order.belongsToMany(Item, { through: 'OrderItem' });
 module.exports = (sequelize, DataTypes) => {
-    return Order;
+  return Order;
 }
-
-item.belongsToMany(orderedItems);
-orderedItems.belongsToMany(item);
