+++ server.js
@@ -6,7 +6,8 @@ const sequelize = require('./models/sequelize.js');
 const users = require('./api/users.js');
 const charges = require('./api/charges.js');
 const sources = require('./api/sources.js');
-const items = require('./api/items.js')
+const items = require('./api/items.js');
+const orders = require('./api/orders.js');
 
 //sync sequelize
 sequelize.sync();
